{bug} - 서버 jwt 작업
작업 : 폴더 네이밍 변경
이유 : 라이브러리 폴더랑 네이밍이 겹치는 문제 발생

#13
