+++ src/docs/docs.go
@@ -36,13 +36,13 @@ const docTemplate = `{
                     "400": {
                         "description": "Bad Request",
                         "schema": {
-                            "$ref": "#/definitions/errorSystem.ResError"
+                            "$ref": "#/definitions/errorCommon.ResError"
                         }
                     },
                     "500": {
                         "description": "Internal Server Error",
                         "schema": {
-                            "$ref": "#/definitions/errorSystem.ResError"
+                            "$ref": "#/definitions/errorCommon.ResError"
                         }
                     }
                 }
@@ -68,13 +68,13 @@ const docTemplate = `{
                     "400": {
                         "description": "Bad Request",
                         "schema": {
-                            "$ref": "#/definitions/errorSystem.ResError"
+                            "$ref": "#/definitions/errorCommon.ResError"
                         }
                     },
                     "500": {
                         "description": "Internal Server Error",
                         "schema": {
-                            "$ref": "#/definitions/errorSystem.ResError"
+                            "$ref": "#/definitions/errorCommon.ResError"
                         }
                     }
                 }
@@ -82,7 +82,7 @@ const docTemplate = `{
         }
     },
     "definitions": {
-        "errorSystem.ResError": {
+        "errorCommon.ResError": {
             "type": "object",
             "properties": {
                 "errType": {
